Add i18n support for en-old / Old English / Englisc
Idiomatic OE uses the same particle ('tha') for 'when' and 'then', occasionally but not always 'tha' for 'when' and 'tha the' for 'then'.

The 7 for 'ond' is as close as a modern typeface gets to the Tironian note for 'et', often used as an abbreviation for 'ond' in OE manuscripts.