Handle interleaved steps and results in the java PrettyFormatter
The java PrettyFormatter has relied on that the call to step() for all
steps in a scenario, come before any call to match() for any step in
the scenario. In Cucumber-JVM version 1.1.3 the call sequence was
changed so this assumption is not true anymore.

To handle interleaved calls to step(), match() and result(), the java
PrettyFormatter is changed to also store the received match and result
data in a list until the end of the scenario. Then the indentations for
the step definition locations are calculated, and the whole scenario is
printed at once.

The final formatted result is not changed, but for long running
scenarios the delay of printing until the end of the scenario will be
notable. Since the printing is deleyed until the end of the scenario,
the printing of the current step as executing is also removed.