Relax the Pretty Formatter spec, executing output not needed
Update the rspec spec for the pretty formatter to explicitly mark the
end of the scenarios with a call to Formatter.eof(). Relax the spec to
allow for but not require the "executing" output for the currently
executing step. The java Pretty Formatter is not printing those, since
it needs to wait to the end of the scenario to be able to calculate the
indentation for the step definition locations for the scenario.